Rename custom target files
