+++ packages/constraint-engine/src/customTargets/testingPlatform/packageDirectory/buildPackageDirectoryReferenceSet.ts
@@ -1,4 +1,4 @@
-import { TestingPlatformPackageDirectorySetTargetPath } from '../packageDirectorySet/buildTestingPlatformPackageDirectorySetReference';
+import { TestingPlatformPackageDirectorySetTargetPath } from '../packageDirectorySet/buildPackageDirectorySetReference';
 import {
   PrefixedTargetPath,
   UnknownTargetPath,
